twitter/requirements.txt
Django==1.11.2
django-bootstrap3==8.2.3
django-friendship==1.5.0
